Rename mint forms
