+++ components/contract-pages/forms/mint/index.tsx
@@ -1,8 +1,8 @@
 import { MarketplaceListForm } from "./list.marketplace";
-import { BundleDropMintForm } from "./mint.bundledrop";
-import { CollectionMintForm } from "./mint.collection";
-import { DropMintForm } from "./mint.drop";
-import { NFTMintForm } from "./mint.nft";
+import { EditionMintForm } from "./mint.edition";
+import { EditionNFTDropMintForm } from "./mint.editiondrop";
+import { NFTCollectionMintForm } from "./mint.nftcollection";
+import { NFTDropMintForm } from "./mint.nftdrop";
 import { TokenMintForm } from "./mint.token";
 import { ProposalMintForm } from "./mint.vote";
 import { DrawerHeader, Heading } from "@chakra-ui/react";
@@ -30,7 +30,7 @@ export const MintForm: React.FC<IMintFormProps> = ({ contract }) => {
         <DrawerHeader>
           <Heading>Mint new NFT</Heading>
         </DrawerHeader>
-        <NFTMintForm contract={contract} />
+        <NFTCollectionMintForm contract={contract} />
       </>
     );
   }
@@ -41,7 +41,7 @@ export const MintForm: React.FC<IMintFormProps> = ({ contract }) => {
         <DrawerHeader>
           <Heading>Create new NFT Drop</Heading>
         </DrawerHeader>
-        <DropMintForm contract={contract} />
+        <NFTDropMintForm contract={contract} />
       </>
     );
   }
@@ -52,7 +52,7 @@ export const MintForm: React.FC<IMintFormProps> = ({ contract }) => {
         <DrawerHeader>
           <Heading>Create new Edition Drop</Heading>
         </DrawerHeader>
-        <BundleDropMintForm contract={contract} />
+        <EditionNFTDropMintForm contract={contract} />
       </>
     );
   }
@@ -63,7 +63,7 @@ export const MintForm: React.FC<IMintFormProps> = ({ contract }) => {
         <DrawerHeader>
           <Heading>Mint new NFT</Heading>
         </DrawerHeader>
-        <CollectionMintForm contract={contract} />
+        <EditionMintForm contract={contract} />
       </>
     );
   }
